Update MaxProductThree
OK, so now that I have considered the negatives, the result has crept up to 66%, there are a few cases which I haven't considered. (Unfortunately, Codility doesn't provide the edge-cases). i.e. in this code, if all negatives in the array it will fail....currently.

The basic idea behind this code is two fold, and kinda works how I would solve the problem just by looking at the array. 

1) you find the largest product of 3 numbers in *one pass* of the array, without sorting it (as no need to)
2) whilst doing this, you make a note of the two largest negative numbers.
3) once the loop has finished, you compare the positive product (X,Y,Z) to the negative product (i.e. largest positive * largest negative * next largest negative). In order to calculate a negative product, you need 2 negative numbers, minimum.
4  you pick either largest positive product, or largest negative product.